Move pages to a separate file
